Add moving average analysis for bypass
This avoids the potential issue that fluctuations in daily flows cause the
wrong amount of gas (and in particular, not enough) to be forwarded to
bypass countries. Uses moving average to prevent a bias towards the most
recent flows for older data (and allows the data to be recreated from
scratch if needed without further introducing this bias).

However, it does mean conservation of mass is not maintained on a daily
basis (but will be maintained over a longer period).